Thu, Jul 21, 2022, 9:12 PM +00:00

+++ readability/readability.c
@@ -17,9 +17,9 @@ int main(void)
     int word_count = count_words(text);
     int sentence_count = count_sentences(text);
 
-    float L = letter_count / word_count * 100;
+    float L = (float) letter_count /  (float) word_count * 100;
     printf ("%f\n", L);
-    float S = sentence_count / word_count * 100;
+    float S = (float) sentence_count / (float) word_count * 100;
     printf ("%f\n", S);
 
     float index = 0.0588 * L - 0.296 * S - 15.8;
